requirements.txt
python-dotenv
streamlit
langchain
huggingface_hub
chromadb
tiktoken
sentence_transformers
pinecone-client
pypdf
replicate